+++ app-web/src/index.tsx
@@ -4,13 +4,16 @@ import ReactDOM from 'react-dom/client'
 import Home from './pages/Home'
 import './styles/global.css'
 import { BackgroundImageProvider } from './stores/HomeBackgroundProvider';
+import { WeatherForecastProvider } from './stores/WeatherForecastProvider';
 
 const root = ReactDOM.createRoot(document.querySelector('#root') as Element);
 
 root.render(
     <React.StrictMode>
-        <BackgroundImageProvider>
-            <Home />
-        </BackgroundImageProvider>
+        <WeatherForecastProvider>
+            <BackgroundImageProvider>
+                <Home />
+            </BackgroundImageProvider>
+        </WeatherForecastProvider>
     </React.StrictMode>
 );