+++ app-web/src/interfaces/api.interface.ts
@@ -1,4 +1,0 @@
-export interface ITemperature {
-    celsius: number,
-    fahrenheit: number
-}