+++ app-web/src/interfaces/context.interface.ts
@@ -1,4 +1,12 @@
+import { ColorTheme } from "./types";
+import { IWeatherForecast } from "./weatherForecast.interface";
+
 export interface IBackgroundImageContextType {
     backgroundImage: string;
     fetchBackgroundImage: () => void;
+}
+
+export interface IWeatherForecastContextType extends IWeatherForecast {
+    fetchWeatherForecast: (city: string) => void;
+    colorTheme: ColorTheme;
 }