+++ app-web/src/interfaces/weatherForecast.interface.ts
@@ -1,0 +1,15 @@
+export interface IWeatherForecast {
+    temperature?: ITemperature,
+    weather?: string,
+    windSpeed?: number,
+    windDirection?: string,
+    humidity?: number,
+    pressure?: number,
+    tomorowTemperature?: ITemperature
+    afterTomorrowTemperature?: ITemperature
+}
+
+export interface ITemperature {
+    celsius: number,
+    fahrenheit: number
+}