+++ app-web/src/utils/convertHextoRgba.ts
@@ -4,7 +4,5 @@ export function hexToRgba(hex: string, opacity: number) {
     const g = (bigint >> 8) & 255;
     const b = bigint & 255;
 
-    console.log(r, g, b)
-
     return `rgba(${r}, ${g}, ${b}, ${opacity})`;
 };