+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.9.0] - 2024-06-21
+
+### Added
+- weather forecast request fully implemented
+- stores and context to get weather forecast
+
 ## [0.8.0] - 2024-06-21
 
 ### Added